Moved rules to correct place, added logging to initialise and fixed an xml doc
